refactor(grovedb): make the terminal non-Merk tree gate a file-versioned function
The gate added in the previous commit was an inline
`if ...terminal_non_merk_tree_child_hash >= 1` branch inside
`prove_subqueries_v1`. That does not follow the versioning system's code
structure: every other gated behaviour in the repo is a module directory with
one file per version and a dispatching `mod.rs`, so each version's behaviour
reads on its own instead of being reconstructed from a conditional.

Restructures it into
`operations/proof/bind_terminal_non_merk_tree/{mod,v0,v1}.rs`, following
`operations/insert/add_element_on_transaction/`:

  - `mod.rs` matches the slot and dispatches, with an `UnknownVersionMismatch`
    arm, and carries the docs on what differs between versions.
  - `v0.rs` is the released no-op. It takes the same arguments and ignores
    them, and documents that the unbound element bytes are a known gap
    preserved because GROVE_V3 is live — not an oversight.
  - `v1.rs` holds the state-root derivation, the self-check and the node
    rewrite. `non_merk_tree_child_hash` moves here from `generate.rs`, since
    only v1 needs it.

The gated unit is the binding step rather than the whole enclosing function,
which keeps the per-version files small — duplicating the ~1000-line
`prove_subqueries_v1` would not have been reasonable. `generate.rs` now just
calls the dispatcher. Passing `&mut Node` and deriving key/value inside also
resolves the borrow that forced the old code to clone them up front.

Behaviour is unchanged: `terminal_non_merk_tree_child_hash_version_gate` still
pins GROVE_V3 to the bare KVValueHash (forgery accepted) and GROVE_V4 to the
child-hash node (forgery rejected). The module is `minimal`-gated, so
verify-only builds are unaffected.
